Issue #96 ビジネスロジック層の実装： ビジネスロジック層の整理と構築、DomainディレクトリにViewModelとModelを分離配置
