+++ iOSEngineerCodeCheck/Domain/ViewModel/ImageLoader.swift
@@ -1,0 +1,40 @@
+//
+//  ImageLoader.swift
+//  iOSEngineerCodeCheck
+//
+//  Created by taro-taryo on 2024/11/10.
+// Copyright © 2024 YUMEMI Inc. All rights reserved.
+// Copyright © 2024 taro-taryo. All rights reserved.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+import Foundation
+import SwiftUI
+
+class ImageLoader: ObservableObject {
+    @Published var image: UIImage?
+    private let imageService: ImageFetchable
+
+    init(imageService: ImageFetchable = ImageService()) {
+        self.imageService = imageService
+    }
+
+    func loadImage(from urlString: String?) {
+        guard let urlString = urlString else { return }
+        imageService.fetchImage(from: urlString) { [weak self] image in
+            DispatchQueue.main.async {
+                self?.image = image
+            }
+        }
+    }
+}
